Move SchemasController to Controllers folder
